refactor(server): split chunkedupload into modular files
